Move snapshot archive generation out of the SnapshotPackagerService
