+++ ledger/Cargo.toml
@@ -43,6 +43,7 @@ solana-sdk = { path = "../sdk", version = "0.23.0" }
 solana-stake-program = { path = "../programs/stake", version = "0.23.0" }
 solana-vote-program = { path = "../programs/vote", version = "0.23.0" }
 sys-info = "0.5.8"
+symlink = "0.1.0"
 tar = "0.4.26"
 thiserror = "1.0"
 tempfile = "3.1.0"
